+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- campaigns page for Cosmere characters
+
 ## [0.4.31] - 2026-04-29
 ### Added
 - selecting ancestry and cultures for Cosmere characters